Added option to update GitVersion using Choco
Added an additional checkbox to the Meta Runner Configuration to allow the running of:

```choco update gitversion.portable```

during the build.  Obviously, this is not something that you woul want to run on each build (as this will add time to the overall build), but from time to time, or when you know there is an update available, a simple modification to the build step can make this happen.